Time: 0 ms (100%), Space: 12.1 MB (66.1%) - LeetHub

+++ 0682-baseball-game/0682-baseball-game.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int calPoints(vector<string>& operations) {
         vector<int> out;
-        int indx = -1;
+        short indx = -1;
         for (string& val : operations)
         {
             if (val == "+")
